move own components into sub-folders for better organization
