+++ Frontend/src/components/general/ListCard/listCard.module.css
