+++ Frontend/src/components/institutionOverview/InstitutionCard/index.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import SettingsIcon from '@mui/icons-material/Settings';
 import Button from '@mui/material/Button';
-import ListCard from 'components/ListCard';
-import ProjectLeaveButton from 'components/ProjectLeaveButton';
-import styles from './projectCard.module.css';
+import ListCard from 'components/general/ListCard';
+import InstitutionLeaveButton from 'components/institutionOverview/InstitutionLeaveButton';
+import styles from './institutionCard.module.css';
 
 import getUser from 'shared/services/mock/user';
 
-function ProjectCard({ project, onLeft }) {
+function InstitutionCard({ institution, onLeft }) {
   const [user, setUser] = useState({});
   useEffect(() => {
     getUser()
@@ -15,12 +15,12 @@ function ProjectCard({ project, onLeft }) {
   }, [setUser]);
 
   const {
-    name, description, adminIds,
-  } = project;
+    name, profilePictureURL, adminIds,
+  } = institution;
   return (
     <ListCard
       title={name}
-      description={description}
+      imageURL={profilePictureURL}
       nextToTitle={(
         <span className={styles.accessRightIndicator}>
           {adminIds.indexOf(user.id) !== -1 ? 'admin' : 'member'}
@@ -37,10 +37,10 @@ function ProjectCard({ project, onLeft }) {
             Settings
           </Button>
         ) : <div key="nothing" />,
-        <ProjectLeaveButton key="leave" project={project} onLeft={onLeft} />,
+        <InstitutionLeaveButton key="leave" institution={institution} onLeft={onLeft} />,
       ]}
     />
   );
 }
 
-export default ProjectCard;
+export default InstitutionCard;
